Fixed `this.$(selector)` returning cached result
Existing usage of `this.$()` will continue to function the same way, while
`this.$(selector)` will work similarly to `this.subject().$(selector)` but
will first ensure that the subject has been rendered.

Also deprecated `this.append()` in favor of its new name, `this.render()` to
help better communicate its purpose.

Fixes #55.