Do not require Ember Data.
Currently, Ember Data is required and installed even if it is not
needed. The majority of functionality in `ember-qunit` is related to
Ember itself, but does not require Ember data.

This removes Ember Data as a runtime dependency, and adds a helpful
error message if it is not found when attempting to use
`moduleForModel`.