Remove `exception` method from QUnit test adapter.
As of Ember 2.17 the `Ember.Test.adapter` is not required to have an
`exception` method, and the method essentially has no purpose. The
default implementation (in the `Ember.Test.QUnitAdapter`) is to
re-throw, but that is _exactly_ what happens by default.

This removes "yet another magical testing thing", and relies on both
Ember and QUnit's default behaviors to ensure that tests properly fail
when unhandled RSVP rejections occur.