Ensure the registry is used for `treeForAddonTestSupport`.
This ensures that the `ember-qunit` files are analyzed by
`ember-auto-import`. `ember-auto-import` requires that the host runs
through the registry (because that is where the analysis actually
happens).

Without this change (e.g. in `ember-qunit@5.0.0-beta.3`) applications
that do **not** have `ember-auto-import` as dependencies themselves _or_
do not have a direct import of `qunit` somewhere in their own code would
simply never have `QUnit` defined.