Relax peer dependency minimum versions to original ember-qunit@5.0.0 versions.
ember-qunit@5.0.0 shipped with `qunit: ^2.13.0` and `@ember/test-helpers: ^2.1.0`.
However, when Dependabot landed a dependency update for our internal
dev dependencies it also increased the minimum for the peerDep (in
https://github.com/emberjs/ember-qunit/commit/719af61effe22013b9d9768d8af9eff9243e27e2
and
https://github.com/emberjs/ember-qunit/commit/fd9fa19ac09a7ad3a3bd88ef2b7b5faa9c9b747c)
both of which were released in ember-qunit@5.1.2.

Changing the minimum version of a peer dependency that we enforce (via
`validatePeerDependencies`) should be considered a breaking change!

This reverts that change and sets the peer dependencies back to its value
when 5.0.0 was published.