Feature: introduce native (ambient) TS types
- Add type definitions and type tests.
- Explicitly specify the SemVer policy in the README.
- Configure CI to run type tests against the package.
- Add a note to the migration docs about how to update to use these.

This will shortly be superceded by types published from source *and*
using the types published by Ember itself. Publishing this ahead of
this provides a path for teams to start using `ember-qunit` with
TypeScript *without* pulling it from DefinitelyTyped, allows us to
remove `ember-qunit` from DefinitelyTyped entirely, and does not block
either of those wins on the timeline of a full TS conversion.

To make this work, we need (because one of our dependencies needs) a
peer and dev dep on `@glimmer/interfaces` and `@glimmer/interfaces`, so
that types which use those will type check. This is: annoying in the
extreme. We will want to keep thinking about how to tackle this as an
ecosystem; for now, they are marked as *optional* so no one will have
things blow up as a result of this at least (and that's also correct:
we *only* need them for TS).