Re-apply version bump to CHANGELOG and package.json
We turned on protection against pushes to `master`, but have not set up
any automation to allow creation of releases via CI, so `release-it`
got through the point where it had published to `npm`, then tried to
push to GitHub, failed, and rolled this back. I'll let it merge, then
tag the result as 6.1.0 for the repo.