Remove pauseTest workaround
Follows-up https://github.com/emberjs/ember-qunit/pull/497 which introduced
this because QUnit 2.8 didn't support changing an existing `assert.timeout()`
by calling it again in the same test (it would leave the old one unchanged,
and start a second timeout).

This was fixed in QUnit 2.9.3, released in Oct 2019.

The ember-qunit package declares a peer dependency on `qunit@2.13.0`,
which should remove the need for this workaround.

Ref https://github.com/emberjs/ember-qunit/issues/496.