Add types for object based test.each() datasets
Prior to this change, we only supported an array for the dataset type
which causes the original `qunit` type definitions to resolve if the
dataset is an object. The fallback will result in the `TestContect`
being removed from the `this` and marking it as `any`.

According the QUnit docs if you pass an object instead of an array to
test.each() the keys of the object become the test label while the
values are the data passed to each run of the test.

This change unions the T[] with Record<string, T> so that both options
can be used as expected.